add: custom repos

+++ src/book/book.controller.ts
@@ -10,33 +10,45 @@ import {
 import { BookService } from './book.service';
 import { CreateBookDto } from './dto/create-book.dto';
 import { UpdateBookDto } from './dto/update-book.dto';
+import { Book } from './entities/book.entity';
 
 @Controller('books')
 export class BookController {
   constructor(private readonly bookService: BookService) {}
 
   @Post()
-  create(@Body() dto: CreateBookDto) {
-    return this.bookService.create(dto);
+  async create(@Body() dto: CreateBookDto): Promise<CreateBookDto> {
+    return await this.bookService.create(dto);
   }
 
   @Get()
-  findAll() {
-    return this.bookService.findAll();
+  async findAll(): Promise<Book[]> {
+    return await this.bookService.findAll();
   }
 
   @Get(':id')
-  findOne(@Param('id') bookId: number) {
-    return this.bookService.findOne(bookId);
+  async findOne(@Param('id') bookId: number): Promise<Book> {
+    return await this.bookService.findOne(bookId);
   }
 
   @Patch(':id')
-  update(@Param('id') bookId: number, @Body() dto: UpdateBookDto) {
-    return this.bookService.update(bookId, dto);
+  async update(
+    @Param('id') bookId: number,
+    @Body() dto: UpdateBookDto,
+  ): Promise<void> {
+    const book = await this.bookService.findOne(bookId);
+    if (!book) {
+      throw { error: 'book is not exist' };
+    }
+    return await this.bookService.update(bookId, dto);
   }
 
   @Delete(':id')
-  remove(@Param('id') bookId: number) {
-    return this.bookService.remove(bookId);
+  async remove(@Param('id') bookId: number): Promise<void> {
+    const book = await this.bookService.findOne(bookId);
+    if (!book) {
+      throw { error: 'book is not exist' };
+    }
+    return await this.bookService.remove(bookId);
   }
 }
